move arm::test_support to acle::test_support
As suggested in the PR comment.
